Refactor: Moved line-specific vector graphics code to line_renderlayer.
The generic vector graphics renderlayer contains code for rendering a generic vector graphics primitive (e.g. polyline, point, etc). The line renderlayer contains specialized shaders and related code for rendering anti-aliased lines.
