+++ src/neuroglancer/sliceview/vector_graphics/frontend.ts
@@ -38,15 +38,13 @@ import {registerSharedObjectOwner, RPC, RpcId, SharedObject} from 'neuroglancer/
 
 const tempMat4 = mat4.create();
 
-export class RenderLayer extends GenericSliceViewRenderLayer {
+export abstract class RenderLayer extends GenericSliceViewRenderLayer {
   sources: VectorGraphicsChunkSource[][];
   shader: ShaderProgram|undefined = undefined;
   shaderUpdated = true;
   rpcId: RpcId|null = null;
   shaderError: WatchableShaderError;
   private sharedObject: SharedObject;
-  private vertexIndexBuffer: Buffer;
-  private normalDirectionBuffer: Buffer;
 
   constructor(
       multiscaleSource: MultiscaleVectorGraphicsChunkSource,
@@ -61,21 +59,6 @@ export class RenderLayer extends GenericSliceViewRenderLayer {
     sharedObject.RPC_TYPE_ID = VECTOR_GRAPHICS_RENDERLAYER_RPC_ID;
     sharedObject.initializeCounterpart(this.chunkManager.rpc!, {'sources': this.sourceIds});
     this.rpcId = sharedObject.rpcId;
-    
-    let vertexIndex = new Float32Array(8);
-    // (1, 0) , (0, 1) , (1, 0) , (0 1)
-    vertexIndex[0] = 1.;
-    vertexIndex[3] = 1.;
-    vertexIndex[4] = 1.;
-    vertexIndex[7] = 1.;
-    this.vertexIndexBuffer = Buffer.fromData(gl, vertexIndex, gl.ARRAY_BUFFER, gl.STATIC_DRAW);
-
-    let normalDirection = new Float32Array(4);
-    normalDirection[0] = 1.;
-    normalDirection[1] = 1.;
-    normalDirection[2] = -1.;
-    normalDirection[3] = -1.;
-    this.normalDirectionBuffer = Buffer.fromData(gl, normalDirection, gl.ARRAY_BUFFER, gl.STATIC_DRAW);
   }
 
   defineShader(builder: ShaderBuilder) {
@@ -83,14 +66,19 @@ export class RenderLayer extends GenericSliceViewRenderLayer {
 void emit(vec4 color) {
   gl_FragColor = color;
 }
+void emitRGBA(vec4 rgba) {
+  emit(vec4(rgba.rgb, rgba.a * uOpacity));
+}
+void emitRGB(vec3 rgb) {
+  emit(vec4(rgb, uOpacity));
+}
+void emitGrayscale(float value) {
+  emit(vec4(value, value, value, uOpacity));
+}
+void emitTransparent() {
+  emit(vec4(0.0, 0.0, 0.0, 0.0));
+}
 `);
-
-    builder.addAttribute('highp float', 'aNormalDirection');
-    builder.addAttribute('highp vec2', 'aVertexIndex');
-
-    builder.addAttribute('highp vec3', 'aVertexFirst');
-    builder.addAttribute('highp vec3', 'aVertexSecond');
-    builder.addUniform('highp mat4', 'uProjection');
   }
 
   beginSlice(_sliceView: SliceView) {
@@ -101,108 +89,21 @@ void emit(vec4 color) {
     return shader;
   }
 
-  endSlice(shader: ShaderProgram) {
-    let gl = this.gl;
+  abstract endSlice(shader: ShaderProgram): void
 
-    gl.disableVertexAttribArray(shader.attribute('aVertexIndex'));
-    gl.disableVertexAttribArray(shader.attribute('aNormalDirection'));
-
-    gl.disableVertexAttribArray(shader.attribute('aVertexFirst'));
-    gl.disableVertexAttribArray(shader.attribute('aVertexSecond'));
-  }
-
-  draw(sliceView: SliceView) {
-    let visibleSources = sliceView.visibleLayers.get(this)!;
-    if (visibleSources.length === 0) {
-      return;
-    }
-
-    this.initializeShader();
-    if (this.shader === undefined) {
-      console.log('error: shader undefined');
-      return;
-    }
-
-    let gl = this.gl;
-
-    let shader = this.beginSlice(sliceView);
-
-    for (let _source of visibleSources) {
-      let source = _source as VectorGraphicsChunkSource;
-      let chunkLayout = source.spec.chunkLayout;
-      let chunks = source.chunks;
-
-      // Compute projection matrix that transforms vertex coordinates to device coordinates
-      gl.uniformMatrix4fv(
-          shader.uniform('uProjection'), false,
-          mat4.multiply(tempMat4, sliceView.dataToDevice, chunkLayout.transform));
-
-      let chunkDataSize: vec3|undefined;
-      let visibleChunks = sliceView.visibleChunks.get(chunkLayout);
-      if (!visibleChunks) {
-        continue;
-      }
-
-      for (let key of visibleChunks) {
-        let chunk = chunks.get(key);
-        if (chunk && chunk.state === ChunkState.GPU_MEMORY) {
-          let numInstances = chunk.numPrimitives;
-
-          this.vertexIndexBuffer.bindToVertexAttrib(
-            shader.attribute('aVertexIndex'),
-            /*components=*/2
-          );
-
-          this.normalDirectionBuffer.bindToVertexAttrib(
-            shader.attribute('aNormalDirection'),
-            /*components=*/1
-          );
-
-          const aVertexFirst = shader.attribute('aVertexFirst');
-          chunk.vertexBuffer.bindToVertexAttrib(
-            aVertexFirst,
-            /*components=*/3,
-            /*attributeType=*/GL_FLOAT,
-            /*normalized=*/false,
-            /*stride=*/6*4,
-            /*offset=*/0
-          );
-          gl.ANGLE_instanced_arrays.vertexAttribDivisorANGLE(aVertexFirst, 1);
-
-          const aVertexSecond = shader.attribute('aVertexSecond');
-          chunk.vertexBuffer.bindToVertexAttrib(
-            aVertexSecond,
-            /*components=*/3,
-            /*attributeType=*/GL_FLOAT,
-            /*normalized=*/false,
-            /*stride=*/6*4,
-            /*offset=*/3*4
-          );
-          gl.ANGLE_instanced_arrays.vertexAttribDivisorANGLE(aVertexSecond, 1);
-
-          gl.ANGLE_instanced_arrays.drawArraysInstancedANGLE(gl.TRIANGLE_STRIP, 0, 4, numInstances);
-
-          gl.ANGLE_instanced_arrays.vertexAttribDivisorANGLE(aVertexFirst, 0);
-          gl.ANGLE_instanced_arrays.vertexAttribDivisorANGLE(aVertexSecond, 0);
-          
-        }
-      }
-    }
-    this.endSlice(shader);
-  }
+  abstract draw(sliceView: SliceView): void 
 }
 
 export class VectorGraphicsChunk extends SliceViewChunk {
   source: VectorGraphicsChunkSource;
   vertexPositions: Float32Array;
   vertexBuffer: Buffer;
-  vertexNormals: Float32Array;
-  numPrimitives: number;
+  numPoints: number;
 
   constructor(source: VectorGraphicsChunkSource, x: any) {
     super(source, x);
     this.vertexPositions = x['vertexPositions'];
-    this.numPrimitives = Math.floor(this.vertexPositions.length / 6); 
+    this.numPoints = Math.floor(this.vertexPositions.length / 3);
   }
 
   copyToGPU(gl: GL) {
